Updated per @khanhnwin feedback
