+++ examples/cookbook/design/themes/README.md
@@ -7,3 +7,7 @@ This recipe explains and demonstrates the following:
 - How to set and apply color and font values to an entire Flutter app.
 - How to override the overall theme's color and font values for
   one or more components in a Flutter app.
+
+## Source Code
+
+https://docs.flutter.dev/cookbook/design/themes
