+++ examples/cookbook/design/themes/lib/main.dart
@@ -14,6 +14,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     const appName = 'Custom Themes';
 
+    // #docregion MaterialApp
     return MaterialApp(
       title: appName,
       theme: ThemeData(
@@ -22,9 +23,11 @@ class MyApp extends StatelessWidget {
         // Define the default brightness and colors.
         colorScheme: ColorScheme.fromSeed(
           seedColor: Colors.purple,
+          // #enddocregion MaterialApp
           // TRY THIS: Change to "Brightness.light"
           //           and see that all colors change
           //           to better contrast a light background.
+          // #docregion MaterialApp
           brightness: Brightness.dark,
         ),
 
@@ -35,10 +38,12 @@ class MyApp extends StatelessWidget {
             fontSize: 72,
             fontWeight: FontWeight.bold,
           ),
+          // #enddocregion MaterialApp
           // TRY THIS: Change one of the GoogleFonts
           //           to "lato", "poppins", or "lora".
           //           The title uses "titleLarge"
           //           and the middle text uses "bodyMedium".
+          // #docregion MaterialApp
           titleLarge: GoogleFonts.oswald(
             fontSize: 30,
             fontStyle: FontStyle.italic,
@@ -51,6 +56,7 @@ class MyApp extends StatelessWidget {
         title: appName,
       ),
     );
+    // #enddocregion MaterialApp
   }
 }
 
@@ -79,9 +85,11 @@ class MyHomePage extends StatelessWidget {
           color: Theme.of(context).colorScheme.primary,
           child: Text(
             'Text with a background color',
+            // #enddocregion Container
             // TRY THIS: Change the Text value
             //           or change the Theme.of(context).textTheme
             //           to "displayLarge" or "displaySmall".
+            // #docregion Container
             style: Theme.of(context).textTheme.bodyMedium!.copyWith(
                   color: Theme.of(context).colorScheme.onPrimary,
                 ),
@@ -104,4 +112,35 @@ class MyHomePage extends StatelessWidget {
       ),
     );
   }
+}
+
+void theme(BuildContext context) {
+  // #docregion Theme
+  Theme(
+    // Create a unique theme with `ThemeData`.
+    data: ThemeData(
+        colorScheme: ColorScheme.fromSeed(
+      seedColor: Colors.pink,
+    )),
+    child: FloatingActionButton(
+      onPressed: () {},
+      child: const Icon(Icons.add),
+    ),
+  );
+  // #enddocregion Theme
+
+  // #docregion ThemeCopyWith
+  Theme(
+    // Find and extend the parent theme using `copyWith`.
+    // To learn more, check out the next section on `Theme.of`.
+    data: Theme.of(context).copyWith(
+        colorScheme: ColorScheme.fromSeed(
+      seedColor: Colors.pink,
+    )),
+    child: const FloatingActionButton(
+      onPressed: null,
+      child: Icon(Icons.add),
+    ),
+  );
+  // #enddocregion ThemeCopyWith
 }
